RockLib.Encryption.XSerializer version 1.0.1
SerializingCrypto can deserialize/decrypt payloads with properties marked with [Encrypt] where the actual property value is not encrypted. This allows for safer deployment strategies when existing properties have been identified as sensitive and in need of encryption. Receiving applications can be deployed first, able to handle both the current plain-text property values and also the encrypted values when sending systems are deployed with the matching change.