RockLib.Encryption version 2.1.0
Changes CompositeCrypto to also implement IAsyncCrypto. Previously, when applied to a CompositeCrypto, the .AsAsync() extension method would always return a SynchronousAsyncCrypto, even if the CompositeCrypto contained cryptos that implement IAsyncCrypto. With the change, CompositeCrypto implements IAsyncCrypto itself and does so by calling .AsAsync() on whatever crypto matches the credential name.